E2503 Refactored Team Class

+++ app/models/team.rb
@@ -1,4 +1,5 @@
 class Team < ApplicationRecord
+  validates :name, uniqueness: { scope: :parent_id, message: "is already in use." }
   has_many :teams_users, dependent: :destroy
   has_many :users, through: :teams_users
   has_many :join_team_requests, dependent: :destroy
@@ -21,33 +22,17 @@ class Team < ApplicationRecord
     Object.const_get(self.name.gsub('Team', '')).find(id)
   end
 
-  # Allowed types of teams -- ASSIGNMENT teams or COURSE teams
-  def self.allowed_types
-    # non-interpolated array of single-quoted strings
-    %w[Assignment Course]
-  end
-
   # Get the participants of the given team
   def participants
     users.where(parent_id: parent_id || current_user_id).flat_map(&:participants)
   end
   alias get_participants participants
 
-  # copies content of one object to the another
+  # # copies content of one object to the another
   def self.copy_content(source, destination)
     source.each do |each_element|
       each_element.copy(destination.id)
     end
-  end
-
-  # enum method for team clone operations
-  def self.team_operation
-    { inherit: 'inherit', bequeath: 'bequeath' }.freeze
-  end
-
-  # Get the response review map
-  def responses
-    participants.flat_map(&:responses)
   end
 
   # Delete the given team
@@ -64,7 +49,7 @@ class Team < ApplicationRecord
   end
 
   # Get the names of the users
-  def author_names
+  def member_names
     names = []
     users.each do |user|
       names << user.fullname
@@ -73,7 +58,7 @@ class Team < ApplicationRecord
   end
 
   # Check if the user exist
-  def user?(user)
+  def is_member?(user)
     users.include? user
   end
 
@@ -82,18 +67,18 @@ class Team < ApplicationRecord
     return false if parent_id.nil? # course team, does not max_team_size
 
     max_team_members = Assignment.find(parent_id).max_team_size
-    curr_team_size = Team.size(id)
+    curr_team_size = size
     curr_team_size >= max_team_members
   end
 
   # Add member to the team, changed to hash by E1776
-  def add_member(user, _assignment_id = nil)
-    raise "The user #{user.name} is already a member of the team #{name}" if user?(user)
+  def add_member(user)
+    raise "The user #{user.name} is already a member of the team #{name}" if is_member?(user)
 
     can_add_member = false
     unless full?
       can_add_member = true
-      t_user = TeamsUser.create(user_id: user.id, team_id: id)
+      t_user = TeamsUser.create!(user_id: user.id, team_id: id)
       parent = TeamNode.find_by(node_object_id: id)
       TeamUserNode.create(parent_id: parent.id, node_object_id: t_user.id)
       add_participant(parent_id, user)
@@ -102,40 +87,15 @@ class Team < ApplicationRecord
     can_add_member
   end
 
-  # Define the size of the team
-  def self.size(team_id)
-    #TeamsUser.where(team_id: team_id).count
-    count = 0
-    members = TeamsUser.where(team_id: team_id)
-    members.each do |member|
-      member_name = member.name
-      unless member_name.include?(' (Mentor)') 
-        count = count + 1
-      end
-    end
-    count
-  end
-
-  # Copy method to copy this team
-  def copy_members(new_team)
-    members = TeamsUser.where(team_id: id)
-    members.each do |member|
-      t_user = TeamsUser.create(team_id: new_team.id, user_id: member.user_id)
-      parent = Object.const_get(parent_entity_type).find(parent_id)
-      TeamUserNode.create(parent_id: parent.id, node_object_id: t_user.id)
-    end
-  end
-
-  # Check if the team exists
-  def self.check_for_existing(parent, name, team_type)
-    list = Object.const_get(team_type + 'Team').where(parent_id: parent.id, name: name)
-    raise TeamExistsError, "The team name #{name} is already in use." unless list.empty?
+  # Returns the number of users in the team
+  def size
+    users.size
   end
 
   # Algorithm
   # Start by adding single members to teams that are one member too small.
   # Add two-member teams to teams that two members too small. etc.
-  def self.randomize_all_by_parent(parent, team_type, min_team_size)
+  def self.create_random_teams(parent, team_type, min_team_size)
     participants = Participant.where(parent_id: parent.id, type: parent.class.to_s + 'Participant', can_mentor: [false, nil])
     participants = participants.sort { rand(-1..1) }
     users = participants.map { |p| User.find(p.user_id) }.to_a
@@ -146,18 +106,29 @@ class Team < ApplicationRecord
         users.delete(User.find(teams_user.user_id))
       end
     end
-    teams.reject! { |team| Team.size(team.id) >= min_team_size }
+    teams.reject! { |team| team.size >= min_team_size }
     # sort teams that still need members by decreasing team size
-    teams.sort_by { |team| Team.size(team.id) }.reverse!
+    teams.sort_by { |team| team.size }.reverse!
     # insert users who are not in any team to teams still need team members
-    assign_single_users_to_teams(min_team_size, parent, teams, users) if !users.empty? && !teams.empty?
+    teams.each do |team|
+      # Calculate how many members this team is missing.
+      member_num_difference = min_team_size - team.size
+      while member_num_difference > 0 && !users.empty?
+        # Add the first user from the list to the team.
+        team.add_member(users.first)
+        # Remove that user from the pool.
+        users.shift
+        member_num_difference -= 1
+      end
+      break if users.empty?
+    end
     # If all the existing teams are fill to the min_team_size and we still have more users, create teams for them.
-    create_team_from_single_users(min_team_size, parent, team_type, users) unless users.empty?
+    team_from_users(min_team_size, parent, team_type, users) unless users.empty?
   end
 
   # Creates teams from a list of users based on minimum team size
   # Then assigns the created team to the parent object
-  def self.create_team_from_single_users(min_team_size, parent, team_type, users)
+  def self.team_from_users(min_team_size, parent, team_type, users)
     num_of_teams = users.length.fdiv(min_team_size).ceil
     next_team_member_index = 0
     (1..num_of_teams).to_a.each do |i|
@@ -167,24 +138,9 @@ class Team < ApplicationRecord
         break if next_team_member_index >= users.length
 
         user = users[next_team_member_index]
-        team.add_member(user, parent.id)
+        team.add_member(user)
         next_team_member_index += 1
       end
-    end
-  end
-
-  # Assigns list of users to list of teams based on minimum team size
-  def self.assign_single_users_to_teams(min_team_size, parent, teams, users)
-    teams.each do |team|
-      curr_team_size = Team.size(team.id)
-      member_num_difference = min_team_size - curr_team_size
-      while member_num_difference > 0
-        team.add_member(users.first, parent.id)
-        users.delete(users.first)
-        member_num_difference -= 1
-        break if users.empty?
-      end
-      break if users.empty?
     end
   end
 
@@ -196,6 +152,18 @@ class Team < ApplicationRecord
     counter = last_team ? last_team.name.scan(/\d+/).first.to_i + 1 : 1
     team_name = "#{_team_name_prefix} Team_#{counter}"
     team_name
+  end
+
+  # E1991 : This method allows us to generate
+  # team names based on whether anonymized view
+  # is set or not. The logic is similar to
+  # existing logic of User model.
+  def name(ip_address = nil)
+    if User.anonymized_view?(ip_address)
+      return "Anonymized_Team_#{self[:id]}"
+    else
+      return self[:name]
+    end
   end
 
   # Extract team members from the csv and push to DB,  changed to hash by E1776
@@ -276,54 +244,29 @@ class Team < ApplicationRecord
   end
 
   # Create the team with corresponding tree node
-  def self.create_team_and_node(id)
-    parent = find_parent_entity id # current_task will be either a course object or an assignment object.
-    team_name = Team.generate_team_name(parent.name)
-    team = create(name: team_name, parent_id: id)
-    # new teamnode will have current_task.id as parent_id and team_id as node_object_id.
-    TeamNode.create(parent_id: id, node_object_id: team.id)
-    ExpertizaLogger.info LoggerMessage.new('Model:Team', '', "New TeamNode created with teamname #{team_name}")
-    team
-  end
+  def self.create_team_and_node(parent_id, user_ids = [])
 
-  # E1991 : This method allows us to generate
-  # team names based on whether anonymized view
-  # is set or not. The logic is similar to
-  # existing logic of User model.
-  def name(ip_address = nil)
-    if User.anonymized_view?(ip_address)
-      return "Anonymized_Team_#{self[:id]}"
-    else
-      return self[:name]
-    end
+    parent = find_parent_entity parent_id # current_task will be either a course object or an assignment object.
+    team_name = Team.generate_team_name(parent.name)
+    team = create(name: team_name, parent_id: parent_id)
+    # new teamnode will have current_task.id as parent_id and team_id as node_object_id.
+    TeamNode.create(parent_id: parent_id, node_object_id: team.id)
+    ExpertizaLogger.info LoggerMessage.new('Model:Team', '', "New TeamNode created with teamname #{team_name}")
+
+    # If user IDs are provided, add them to the team.
+    user_ids.each do |user_id|
+      team_user = TeamsUser.where(user_id: user_id)
+                           .find { |tu| tu.team.parent_id == parent_id }
+      team_user.destroy if team_user
+      team.add_member(User.find(user_id))
+    end unless user_ids.empty?
+
+    team
   end
 
   # REFACTOR END:: class methods import export moved from course_team & assignment_team to here
 
-  # Create the team with corresponding tree node and given users
-  def self.create_team_with_users(parent_id, user_ids)
-    team = create_team_and_node(parent_id)
-
-    user_ids.each do |user_id|
-      remove_user_from_previous_team(parent_id, user_id)
-
-      # Create new team_user and team_user node
-      team.add_member(User.find(user_id))
-    end
-    team
-  end
-
-  # Removes the specified user from any team of the specified assignment
-  def self.remove_user_from_previous_team(parent_id, user_id)
-    team_user = TeamsUser.where(user_id: user_id).find { |team_user_obj| team_user_obj.team.parent_id == parent_id }
-    begin
-      team_user.destroy
-    rescue StandardError
-      nil
-    end
-  end
-
-  def self.find_team_users(assignment_id, user_id)
+  def self.find_team_for_user(assignment_id, user_id)
     TeamsUser.joins('INNER JOIN teams ON teams_users.team_id = teams.id')
              .select('teams.id as t_id')
              .where('teams.parent_id = ? and teams_users.user_id = ?', assignment_id, user_id)
@@ -334,8 +277,8 @@ class Team < ApplicationRecord
     participants.include?(participant)
   end
 
-  # Get team's full name
-  def fullname
-    name
-  end
+  # Get team's full name from legacy codebase that is no longer used
+  # def fullname
+  #   name
+  # end
 end
